Added manual "insert reference" button
This plugin now features an extra button in the Insert menu to manually
insert a reference, without having to type one of the triggering
sentences. Added with some translations. Reference is inserted in last
known selected position in the editor. Will need to ask if insertion
should be limited to a certain part of the document.